Move code from RequestOld. Fix tests.
