+++ AndroidSDKCore/src/main/java/com/leanplum/Leanplum.java
@@ -46,6 +46,7 @@ import com.leanplum.internal.Log;
 import com.leanplum.internal.OperationQueue;
 import com.leanplum.internal.Registration;
 import com.leanplum.internal.RequestOld;
+import com.leanplum.internal.RequestOldUtil;
 import com.leanplum.internal.RequestSender;
 import com.leanplum.internal.Util;
 import com.leanplum.internal.Util.DeviceIdInfo;
@@ -718,7 +719,7 @@ public class Leanplum {
   }
 
   private static void handleStartResponse(final JSONObject response) {
-    boolean success = RequestOld.isResponseSuccess(response);
+    boolean success = RequestOldUtil.isResponseSuccess(response);
     Leanplum.countAggregator().incrementCount("on_start_response");
     if (!success) {
       try {
